requirements.txt
timm==0.5.4
hydra-core==1.2.0
# git+https://github.com/openai/CLIP.git
ftfy==6.1.1
tqdm==4.64.0
regex==2019.11.1
point-transformer-pytorch==0.1.5
sentence_transformers==2.2.2
numpy-quaternion==2022.4.2
# git+https://github.com/facebookresearch/detectron2.git
pandas==1.4.3
pyntcloud==0.3.1
hydra-submitit-launcher==1.2.0
torch-encoding==1.2.2b20200804
torchmetrics==0.6.0
sentence-transformers==2.2.2
opencv-python==4.5.5.64
imageio==2.19.3
altair==4.2.0
streamlit==1.12.2
protobuf==3.20.1
matplotlib==3.5.2
test-tube==0.7.5
wandb==0.13.3
omegaconf==2.2.2
numpy==1.21.4
einops==0.4.1
pytorch-lightning==1.3.5
pyquaternion==0.9.9
pyliblzfse==0.4.1
open3d==0.15.2